Update naming and documentation.
Empty is a really bad name because it is very imprecise. It could mean
an empty set of fields, it could mean that the fields have empty types,
a very different property. Moreover in Rust there is a difference between
struct Foo{}, struct Foo(), and struct Foo, all of which have "empty" set
of fields, but will be represented differently.